requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Celery==5.3.4
redis==5.0.1
Pillow>=10.0.1
ReportLab==4.0.7
Werkzeug==2.3.7
